+++ src/components/Filters/FilterItem/FilterItem.module.css
@@ -1,26 +1,37 @@
 .filterButton {
-    font-style: normal;
-    font-weight: 400;
-    font-size: 16px;
-    line-height: 24px;
-    border: 1px solid #ffffff;
-    border-radius: 60px;
-    padding: 6px 20px;
-  }
+  font-style: normal;
+  font-weight: 400;
+  font-size: 16px;
+  line-height: 24px;
+  border: 1px solid #ffffff;
+  border-radius: 60px;
+  padding: 6px 20px;
+}
 
-  
-  .filterButton:not(:last-child) {
-    margin-right: 10px;
-  }
-  
-  ._btnText:hover {
-    border-color: #d9b6ff;
-    color: #d9b6ff;
-    cursor: pointer;
-  }
+.filterButton:not(:last-child) {
+  margin-right: 10px;
+}
 
-  ._btnText:active {
-    border-color: #ad61ff;
-    color: #ad61ff;
-    cursor: pointer;
-  }
+._btnText:hover {
+  border-color: #d9b6ff;
+  color: #d9b6ff;
+  cursor: pointer;
+}
+
+._btnText:active {
+  border-color: #ad61ff;
+  color: #ad61ff;
+  cursor: pointer;
+}
+
+.filterListDown {
+  position: absolute;
+  width: 248px;
+  height: 305px;
+  padding: 34px;
+  background: #313131;
+  border-radius: 12px;
+  gap: 10px;
+  left: 386px;
+  top: 288px;
+}
